backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
fastf1==3.5.3
redis==5.2.1
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.3
anthropic==0.40.0
